+++ app/controllers/settings/base_controller.rb
@@ -1,0 +1,5 @@
+module Settings
+  class BaseController < ApplicationController
+    layout "application"
+  end
+end