+++ db/seeds.rb
@@ -1,1 +1,2 @@
 User.ensure_local_account! if User.local_auth_configured?
+SiteSetting.instance